Clean up MenuButtonManager
